requirements.txt
requests
beautifulsoup4
discord.py
pymongo
python-dotenv